Initial commit. Faster FOV calculations.
Greatly reduces the delay caused by FOV-related updates in the game.
Reduces stuttering while moving, after each shot, and explosion.

Done by allowing partial visibility updates of tiles and units for cases
when we know an event has occurred and the exact tiles affected. This
creates a small circle sector originating at the observing unit with its
arc width set by the radius of the event. Any tiles and/or units in this
area may then be updated, which is faster than updating everything
within that unit's full vision sector.

Also changed the way the unit visibility algorithm scales, from an
area-based algorithm to be based on the number of units in range. This
scales better with view distance.