Improved quick search in Inventory (by Stian)
The main difference is that it searches within an items' categories and loaded ammo for matches as well. Thereby it is possible to look up all items of a certain damage type, weapon type etc. I only enabled this extra feature for researched items though. It also skips searching by name on alien artifacts to reduce spoilers.