Interception Window UI Enhancement
Right now, in the interception window if we right-click a craft that is in the air, it will center on the globe on that craft, but if we right click a craft at base, it does nothing.   Would it be possible to have a right-click of the docked craft open up the equip-craft window in the correct base?

https://openxcom.org/forum/index.php/topic,7706.0.html