Prevent AI frozen in place
I received a report of AI freeze, which looks like this:
1. AI runs setupEscape() and decides it should escape to the same tile
as it currently occupies (for whatever reason); _escapeTUs is set to 1
2. AI evaluation sets _AIMode to 3 (=escape)
3. at the end of the think() cycle, _escapeTUs is not reset
(because there is no need to move)

Starting from this point setupEscape() is not executed anymore,
because _escapeTUs != 0. And as long as the enemy sees (or knows about)
xcom, it will not re-evaluate. Thus every turn it will just stay where
it is without doing anything (only thing it can do are reaction shots).

The freeze breaks only after turn 20 (or other forced re-evaluate).
Or after save and reload... because _escapeTUs is not saved
and will get reset back to 0 after reload.

# Conflicts:
#	src/Savegame/SavedBattleGame.cpp