Fixed order-dependent facility upgrading
History from Discord to search for (2020-04-24):

Ohartenstein: They count as "fully there" for everything but:
1. being able to dismantle them
2. building off of them
3. using their functionality

#2 has been changed in this commit... it is now allowed.
It also automatically fixes the order-dependent facility upgrading.

More comments from Discord:

Yankes: but this was not for normal build que? Because in case of
upgrading you already built a building there and you should be allowed
to expand further

Meridian: I'd be willing to exchange that (potential) misuse for having
order-independent upgrades